pyproject.toml
[tool.poetry]
name = "python-ab2-hw-1-k8s-ilystsov"
version = "0.1.0"
description = ""
authors = ["ilya <lystsov.ilya@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.2"
uvicorn = "^0.27.1"
psycopg2-binary = "^2.9.9"
sqlalchemy = "^2.0.27"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.3"
python-multipart = "^0.0.9"
prometheus-fastapi-instrumentator = "^6.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
